compiling

+++ hw3/vm.c
@@ -223,7 +223,7 @@ allocuvm(pde_t *pgdir, uint oldsz, uint newsz)
   char *mem;
   uint a;
 
-  if((newsz >= KERNBASE)||( newsz >= (p2v(PHYSTOP) - proc->ssm)))
+  if((newsz >= KERNBASE)||( newsz >= (uint)(p2v(PHYSTOP) - proc->ssm)))
     return 0;
   if(newsz < oldsz)
     return oldsz;
